+++ service-accounts/schema/purge.sql
@@ -1,0 +1,6 @@
+drop table if exists  account_platform_data;
+
+drop table if exists  account_platforms;
+
+drop table if exists accounts;
+
